# Refactor PlatformManager for performance