@me4502 2022-07-23
Refactors for performance that make the code substantially messier (the fact you didn't follow the style guide at all aside), without any proof of speed-up or evidence that it's necessary, won't be accepted.

Provide a microbenchmark and CPU profile that show why these changes make sense to do if you want to PR something like this.

I'm also not convinced it couldn't be done in a substantially cleaner manner, if it's true that this has a performance impact