@anonyco 2022-07-23
I have significantly simplified these changes to the bare minimum: the removal of just the `synchronized`.

1. I was unaware that there was a style guide. I read all the markdown and txt files, including CONTRIBUTING.md. Could you refer me to where to find the style guide?

2. The results of my microbenchmark are below:

```
Before:
     Calculated BlockType.getRichName x10000 in 84128654ns
     Calculated BlockType.getRichName x10000 in 14269327ns
     Calculated BlockType.getRichName x10000 in 7806132ns

After:
     Calculated BlockType.getRichName x10000 in 60178043ns
     Calculated BlockType.getRichName x10000 in 7953745ns
     Calculated BlockType.getRichName x10000 in 4852244ns
```

Across the runs above, we see 40%-79% better performance for `BlockType.getRichName` with this small change.

3. I am very unfamiliar with the available Java tools and do not know "CPU profile." I am not asking you to waste your time explaining every step to me, rather I would really appreciate it if you could give me a piece of text or a name of a software/tool that I could look up and research to learn more about what the common way to do this is in the WorldEdit project.

Many thanks for your time and attention to this.