@me4502 2024-07-07
closing in favour of https://github.com/EngineHub/WorldEdit/pull/2571